refactor(data-model): `JSON_CODEC` moves to `codec-interface`
It was in the top-level `interface.ts` because at the time there was nowhere
better: everything under `codec-*` had importers running the other way, and a
symbol in the dependency-free type universe was reachable from anywhere
without closing a cycle. Its own doc comment said so.

`codec-interface/` is that place now. It imports no runtime value from
anywhere in the package, so nothing taken from it can close a cycle either --
and it is where `CODEC` already lives, which is the symbol this one is the
per-format counterpart to. The paragraph justifying the old home goes with the
move, having become an explanation of something that is no longer true.

Fourteen files repointed. The type universe is left declaring types and the
abstract root, which is what it is for.
