+++ packages/data-model/src/codec-interface/index.ts
@@ -3,6 +3,7 @@ export {
   type CodecForFormat,
   type FabricClassWithNonterminalCodec,
   type FabricCodec,
+  JSON_CODEC,
   type NonterminalCodec,
   type ReconstructionContext,
   type SerializationContext,
