+++ packages/data-model/test/fabric-primitives/FabricBytes.test.ts
@@ -11,7 +11,7 @@
  * offset is at or past the end copies nothing rather than failing.
  */
 
-import { JSON_CODEC } from "@/interface.ts";
+import { JSON_CODEC } from "@/codec-interface/interface.ts";
 import { describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
