+++ packages/data-model/test/fabric-primitives/FabricHash.test.ts
@@ -11,7 +11,7 @@
  * when a caller explicitly transfers it.
  */
 
-import { JSON_CODEC } from "@/interface.ts";
+import { JSON_CODEC } from "@/codec-interface/interface.ts";
 import { describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
