+++ packages/data-model/test/fabric-primitives/FabricRegExp.test.ts
@@ -15,7 +15,7 @@
  * as well -- two values differing only in it hash differently.
  */
 
-import { JSON_CODEC } from "@/interface.ts";
+import { JSON_CODEC } from "@/codec-interface/interface.ts";
 import { describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
